+++ app/lib/ui/components/form_input/password_input/password_input_demo_screen.dart
@@ -45,7 +45,7 @@ class PasswordInputDemoScreen extends StatefulWidget {
 
 class _PasswordInputDemoScreenState extends State<PasswordInputDemoScreen> {
   final _scaffoldKey = GlobalKey<ScaffoldState>();
-  bool _isBottomSheetExpanded = false;
+  bool _isBottomSheetExpanded = true;
 
   void _onExpansionChanged(bool isExpanded) {
     setState(() {