+++ ouds_core/lib/components/checkbox/ouds_checkbox.dart
@@ -62,6 +62,7 @@ enum ToggleableState { off, indeterminate, on }
 ///       });
 ///     },
 ///   isError: false,
+///   readOnly: false,
 /// );
 /// ```
 ///
@@ -70,6 +71,7 @@ class OudsCheckbox extends StatefulWidget {
   final ValueChanged<bool?>? onChanged;
   final bool isError;
   final bool tristate;
+  final bool readOnly;
 
   const OudsCheckbox({
     super.key,
@@ -77,6 +79,7 @@ class OudsCheckbox extends StatefulWidget {
     required this.onChanged,
     this.isError = false,
     this.tristate = false,
+    this.readOnly = false,
   });
 
   @override
@@ -102,13 +105,15 @@ class _OudsCheckboxState extends State<OudsCheckbox> {
   Widget build(BuildContext context) {
     final interactionModelHover = OudsInheritedInteractionModel.of(context, InteractionAspect.hover);
     final interactionModelPressed = OudsInheritedInteractionModel.of(context, InteractionAspect.pressed);
-    final isHovered = interactionModelHover?.state.isHovered ?? false;
-    final isPressed = interactionModelPressed?.state.isPressed ?? false;
+    final isHoveredInherited = interactionModelHover?.state.isHovered ?? false;
+    final isPressedInherited = interactionModelPressed?.state.isPressed ?? false;
+    final bool isReadOnly = widget.readOnly;
 
     final checkboxStateDeterminer = OudsControlStateDeterminer(
-      enabled: widget.onChanged != null,
-      isPressed: isPressed || _isPressed,
-      isHovered: isHovered || _isHovered,
+      enabled: (widget.onChanged != null) || isReadOnly,
+      isPressed: (!isReadOnly) && (isPressedInherited || _isPressed),
+      isHovered: (!isReadOnly) && (isHoveredInherited || _isHovered),
+      isReadOnly: isReadOnly,
     );
 
     final checkboxState = checkboxStateDeterminer.determineControlState();
@@ -116,6 +121,7 @@ class _OudsCheckboxState extends State<OudsCheckbox> {
     final checkboxBackgroundModifier = OudsControlBackgroundModifier(context);
     final checkboxTickModifier = OudsControlTickModifier(context);
     final checkbox = OudsTheme.of(context).componentsTokens(context).checkbox;
+    final controlItem = OudsTheme.of(context).componentsTokens(context).controlItem;
     final l10n = OudsLocalizations.of(context);
 
     String? semanticsLabel = widget.value == true
@@ -130,7 +136,7 @@ class _OudsCheckboxState extends State<OudsCheckbox> {
     }
 
     return Semantics(
-      enabled: widget.onChanged != null,
+      enabled: widget.onChanged != null && !(widget.readOnly),
       value: semanticsLabel,
       label: widget.tristate == true ? l10n?.core_checkbox_indeterminateCheckbox_a11y : l10n?.core_checkbox_checkbox_a11y,
       hint: widget.isError ? l10n?.core_common_onError_a11y : null,
@@ -139,7 +145,7 @@ class _OudsCheckboxState extends State<OudsCheckbox> {
         child: SizedBox(
           width: checkbox.sizeMaxHeight,
           child: InkWell(
-            onTap: widget.onChanged != null
+            onTap: (!isReadOnly && widget.onChanged != null)
                 ? () {
                     _isPressed = true;
                     SchedulerBinding.instance.addPostFrameCallback((_) {
@@ -182,7 +188,7 @@ class _OudsCheckboxState extends State<OudsCheckbox> {
               decoration: BoxDecoration(
                 color: _isPressed ? checkboxBackgroundModifier.getBackgroundColor(checkboxState) : Colors.transparent,
                 borderRadius: BorderRadius.circular(
-                  checkboxBorderModifier.getBorderRadius(checkbox),
+                  checkboxBorderModifier.getBorderRadius(controlItem.borderRadiusItemOnly),
                 ),
               ),
               child: Center(
@@ -192,7 +198,7 @@ class _OudsCheckboxState extends State<OudsCheckbox> {
                     height: checkbox.sizeIndicator,
                     child: ClipRRect(
                       borderRadius: BorderRadius.circular(
-                        checkboxBorderModifier.getBorderRadius(checkbox),
+                        checkboxBorderModifier.getBorderRadius(checkbox.borderRadius),
                       ),
                       child: Stack(
                         fit: StackFit.expand,
@@ -213,7 +219,7 @@ class _OudsCheckboxState extends State<OudsCheckbox> {
                                 ),
                               ),
                               borderRadius: BorderRadius.circular(
-                                checkboxBorderModifier.getBorderRadius(checkbox),
+                                checkboxBorderModifier.getBorderRadius(checkbox.borderRadius),
                               ),
                             ),
                           ),